+++ .github/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## 20210327
+
+* Version 1.1.0
+* Added `New Voter Request` function
+* Added error handling with Angular SnackBar
+* Added new event capabilities: `Delegate` and `New Voter Request`
+* Improved UX for `Explore` tab with more feedback about events
+* Updated `About` and `README` to explain new onboarding procedures
+
 ## 20210314
 
 First commit